Fixed distance_to_point, fixed friction, fixed collision, fixed android audio, added support for working_directory and mouse_button. Fixed missing event segfaults. Implemented bad tile system and destroy event
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@814 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL/GSblend.h
@@ -54,4 +54,4 @@ const int bm_inv_dest_color   =10; ///GL_ONE_MINUS_DST_COLOR    //only for src
 const int bm_src_alpha_sat    =11; ///GL_SRC_ALPHA_SATURATE     //only for src
 
 int draw_set_blend_mode(int mode);
-int draw_set_blend_mode_ext(int src,int dest);
+int draw_set_blend_mode_ext(double src,double dest);
